Fixed Properties
- Changed Timeout in options to be a float rather than an int
- Added missing IpAddress in response
- Added missing IpAddress checks in tests

+++ MCStatus.IO.NET/MCStatus.IO.NET/Models/StatusOptions.cs
@@ -3,7 +3,7 @@
 public abstract class StatusOptions
 {
   public string? BaseUrl { get; set; }
-  public int? Timeout { get; set; }
+  public float? Timeout { get; set; }
 
   public virtual string AsQueryParams()
   {
